feat(frontend): move list pages to SSR prefetch
Tasks, Test Runs, Test Sets and Tests pages become async server components that prefetch page 1 via prefetchList and hand initialData to new client components, following the metrics shape. The page-level gate blocks collapse into useListAuthGate. Tests skips the prefetch when an Insights deep link is present, since its test-id filter resolves client-side. Page tests move next to the client components they render.
