+++ apps/frontend/src/app/(protected)/tests/components/TestsPageClient.tsx
@@ -1,0 +1,217 @@
+'use client';
+
+import * as React from 'react';
+import { useRouter, useSearchParams } from 'next/navigation';
+import Box from '@mui/material/Box';
+import { useQueryClient } from '@tanstack/react-query';
+import EditNoteIcon from '@mui/icons-material/EditNote';
+import DownloadOutlinedIcon from '@mui/icons-material/DownloadOutlined';
+import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline';
+import { PageLayout } from '@/components/layout/PageLayout';
+import { Fab, FabGroup } from '@/components/common/Fab';
+import { CategoryIcon } from '@/components/icons';
+import TestsGrid, { type TestsBulkActionsState } from './TestsGrid';
+import FileImportDrawer from '@/app/(protected)/test-sets/components/FileImportDrawer';
+import { testsList } from './list';
+import { useDocumentTitle } from '@/hooks/useDocumentTitle';
+import { useOnboarding } from '@/contexts/OnboardingContext';
+import { parseInsightsFailedTestsSearchParams } from '@/app/(protected)/insights/utils/insights-failed-tests';
+import { useEndpoint } from '@/hooks/useEndpoints';
+import { Can, useCan } from '@/components/common/Can';
+import { Capability } from '@/constants/capabilities';
+import { useListAuthGate } from '@/hooks/useListAuthGate';
+import { useNotifications } from '@/components/common/NotificationContext';
+import { testKeys, testSetKeys } from '@/constants/query-keys';
+import type { TestDetail } from '@/utils/api-client/interfaces/tests';
+
+interface TestsPageClientProps {
+  /**
+   * Server-fetched first page — when present, skips the initial client fetch.
+   * Absent when an Insights deep link is active: its test-id filter resolves
+   * client-side, so the server can't build the first page's `$filter`.
+   */
+  initialData?: TestDetail[];
+  initialTotalCount?: number;
+}
+
+export default function TestsPageClient({
+  initialData,
+  initialTotalCount = 0,
+}: TestsPageClientProps) {
+  const router = useRouter();
+  const searchParams = useSearchParams();
+  const queryClient = useQueryClient();
+  const notifications = useNotifications();
+  const [fileImportDrawerOpen, setFileImportDrawerOpen] = React.useState(false);
+  const [bulkActions, setBulkActions] = React.useState<
+    Pick<TestsBulkActionsState, 'visible' | 'assignDisabled'>
+  >({ visible: false, assignDisabled: false });
+  const bulkHandlersRef = React.useRef<
+    Pick<TestsBulkActionsState, 'onAssign' | 'onDelete'>
+  >({
+    onAssign: () => {},
+    onDelete: () => {},
+  });
+
+  const handleBulkActionsChange = React.useCallback(
+    (actions: TestsBulkActionsState) => {
+      setBulkActions({
+        visible: actions.visible,
+        assignDisabled: actions.assignDisabled,
+      });
+      bulkHandlersRef.current = {
+        onAssign: actions.onAssign,
+        onDelete: actions.onDelete,
+      };
+    },
+    []
+  );
+  const { activeTour, startTour } = useOnboarding();
+  const gate = useListAuthGate(testsList);
+  const canCreate = useCan(Capability.Test.CREATE);
+
+  const insightsFailedFilter = React.useMemo(
+    () =>
+      searchParams ? parseInsightsFailedTestsSearchParams(searchParams) : null,
+    [searchParams]
+  );
+  const { data: insightsEndpoint } = useEndpoint(
+    insightsFailedFilter?.endpointId ?? '',
+    !!insightsFailedFilter
+  );
+  const insightsEndpointName = insightsEndpoint?.name;
+
+  useDocumentTitle('Tests');
+
+  const tourParam = searchParams?.get('tour');
+  const isOnTestCasesTour =
+    tourParam === 'testCases' || activeTour === 'testCases';
+  const shouldDisableAddButton = activeTour !== null && !isOnTestCasesTour;
+
+  React.useEffect(() => {
+    if (tourParam === 'testCases') {
+      const timeout = setTimeout(() => {
+        startTour('testCases');
+      }, 300);
+      return () => clearTimeout(timeout);
+    }
+  }, [tourParam, startTour]);
+
+  React.useEffect(() => {
+    const openGeneration = searchParams?.get('openGeneration');
+    if (openGeneration === 'true') {
+      router.push('/test-sets/new-generated');
+      const newUrl = new URL(window.location.href);
+      newUrl.searchParams.delete('openGeneration');
+      window.history.replaceState({}, '', newUrl.toString());
+    }
+  }, [searchParams, router]);
+
+  const handleCreateManual = React.useCallback(() => {
+    if (activeTour === 'testCases') return;
+    router.push('/tests/new-manual');
+  }, [activeTour, router]);
+
+  const handleFileImportSuccess = React.useCallback(
+    (testSetId: string) => {
+      setFileImportDrawerOpen(false);
+      queryClient.invalidateQueries({ queryKey: testKeys.all() });
+      queryClient.invalidateQueries({ queryKey: testSetKeys.all() });
+      notifications.show('Tests imported successfully', {
+        severity: 'success',
+      });
+      router.push(`/test-sets/${testSetId}`);
+    },
+    [queryClient, notifications, router]
+  );
+
+  React.useEffect(() => {
+    const handleTourOpenModal = () => {
+      router.push('/test-sets/new-generated');
+    };
+    window.addEventListener('tour-open-test-modal', handleTourOpenModal);
+    return () => {
+      window.removeEventListener('tour-open-test-modal', handleTourOpenModal);
+    };
+  }, [router]);
+
+  if (!gate.ready) return gate.node;
+
+  return (
+    <>
+      <PageLayout
+        title="Tests"
+        description="Individual test cases that evaluate your AI endpoints for quality, safety, and reliability."
+        breadcrumbs={[]}
+        actions={
+          <FabGroup>
+            {bulkActions.visible && (
+              <>
+                <Can capability={Capability.TestSet.UPDATE}>
+                  <Fab
+                    icon={<CategoryIcon sx={{ fontSize: 28 }} />}
+                    tooltip={
+                      bulkActions.assignDisabled
+                        ? 'Select tests with the same test type'
+                        : 'Assign to Test Set'
+                    }
+                    aria-label="Assign to Test Set"
+                    onClick={() => bulkHandlersRef.current.onAssign()}
+                    disabled={bulkActions.assignDisabled}
+                  />
+                </Can>
+                <Can capability={Capability.Test.DELETE}>
+                  <Fab
+                    icon={<DeleteOutlineIcon sx={{ fontSize: 28 }} />}
+                    tooltip="Delete Tests"
+                    aria-label="Delete Tests"
+                    onClick={() => bulkHandlersRef.current.onDelete()}
+                    sx={{
+                      bgcolor: 'error.main',
+                      '&:hover': { bgcolor: 'error.dark' },
+                    }}
+                  />
+                </Can>
+              </>
+            )}
+            <Can capability={Capability.File.IMPORT}>
+              <Fab
+                icon={<DownloadOutlinedIcon />}
+                tooltip="Import tests"
+                onClick={() => setFileImportDrawerOpen(true)}
+              />
+            </Can>
+            <Can capability={Capability.Test.CREATE}>
+              <Fab
+                icon={<EditNoteIcon />}
+                tooltip="Manual test"
+                aria-label="Manual test"
+                onClick={handleCreateManual}
+                disabled={shouldDisableAddButton}
+              />
+            </Can>
+          </FabGroup>
+        }
+      >
+        <Box sx={{ mt: 2, mb: 2 }}>
+          <TestsGrid
+            onNewTest={handleCreateManual}
+            disableAddButton={shouldDisableAddButton}
+            canCreate={canCreate}
+            insightsFailedFilter={insightsFailedFilter}
+            insightsEndpointName={insightsEndpointName}
+            onBulkActionsChange={handleBulkActionsChange}
+            initialData={initialData}
+            initialTotalCount={initialTotalCount}
+          />
+        </Box>
+      </PageLayout>
+
+      <FileImportDrawer
+        open={fileImportDrawerOpen}
+        onClose={() => setFileImportDrawerOpen(false)}
+        onSuccess={handleFileImportSuccess}
+      />
+    </>
+  );
+}
